Add checkbox for selecting draft recipe.
When a recipe is a draft only title is required

+++ lib/maria/recipes/recipe.ex
@@ -24,8 +24,9 @@ defmodule Maria.Recipes.Recipe do
   @doc false
   def changeset(recipe, attrs) do
     recipe
-    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :link, :tags, :user_id])
-    |> validate_required([:title, :description, :directions, :mins, :ingredients, :tags])
+    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :link, :tags, :is_draft, :user_id])
+    |> validate_required([:title])
+    |> validate_required_not_draft([:description, :directions, :mins, :ingredients, :tags])
     |> validate_cover(:required)
     |> unique_constraint(:title)
     |> foreign_key_constraint(:user_id)
@@ -34,8 +35,9 @@ defmodule Maria.Recipes.Recipe do
   @doc false
   def changeset_update(recipe, attrs) do
     recipe
-    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :link, :tags, :editor_id])
-    |> validate_required([:title, :description, :directions, :mins, :ingredients, :tags])
+    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :link, :tags, :is_draft, :editor_id])
+    |> validate_required([:title])
+    |> validate_required_not_draft([:description, :directions, :mins, :ingredients, :tags])
     |> validate_cover(:not_required)
     |> foreign_key_constraint(:editor_id)
   end
@@ -43,9 +45,19 @@ defmodule Maria.Recipes.Recipe do
   @doc false
   def changeset_validate(recipe, attrs) do
     recipe
-    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :tags])
-    |> validate_required([:title, :description, :directions, :mins, :ingredients, :tags])
+    |> cast(attrs, [:title, :description, :directions, :mins, :ingredients, :yield, :tags, :is_draft])
+    |> validate_required([:title])
+    |> validate_required_not_draft([:description, :directions, :mins, :ingredients, :tags])
     |> foreign_key_constraint(:user_id)
+  end
+
+  defp validate_required_not_draft(changeset, attrs) do
+    if get_field(changeset, :is_draft) do
+      changeset
+    else
+      changeset
+      |> validate_required(attrs)
+    end
   end
 
   @rx  ~r/(?i)\.(jpg|jpeg|png|gif)$/
@@ -67,7 +79,9 @@ defmodule Maria.Recipes.Recipe do
   end
 
   def delete_cover(recipe) do
-    MariaWeb.File.delete(recipe.cover)
+    if recipe.cover do
+      MariaWeb.File.delete(recipe.cover)
+    end
 
     recipe
   end
